requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pydantic
pydantic-settings
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
pillow
qrcode
python-slugify
httpx
celery
prometheus-client
python-i18n
Babel

asyncpg